Separate server setup and routing, simplified tests using supertest
